--- a/go.mod
+++ b/go.mod
@@ -4,19 +4,18 @@ go 1.18
 
 require (
 	github.com/golang/protobuf v1.5.2
-	github.com/google/cel-go v0.11.2 //v0.6.0
+	github.com/google/cel-go v0.13.0 //v0.6.0
 	github.com/jedib0t/go-pretty/v6 v6.2.2
 	github.com/matryer/is v1.4.0
 	github.com/mattn/go-sqlite3 v1.14.7
-	google.golang.org/genproto v0.0.0-20220310185008-1973136f34c6
-	google.golang.org/protobuf v1.27.1
+	google.golang.org/genproto v0.0.0-20221027153422-115e99e71e1c
+	google.golang.org/protobuf v1.28.1
 )
 
 require (
-	github.com/antlr/antlr4/runtime/Go/antlr v0.0.0-20220209173558-ad29539cd2e9 // indirect
+	github.com/antlr/antlr4/runtime/Go/antlr v1.4.10 // indirect
 	github.com/mattn/go-runewidth v0.0.9 // indirect
 	github.com/stoewer/go-strcase v1.2.0 // indirect
-	golang.org/x/sys v0.0.0-20210510120138-977fb7262007 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
+	golang.org/x/sys v0.0.0-20220728004956-3c1f35247d10 // indirect
+	golang.org/x/text v0.3.8 // indirect
 )