--- a/go.mod
+++ b/go.mod
@@ -5,7 +5,7 @@ go 1.18
 require (
 	github.com/golang/protobuf v1.5.2
 	github.com/google/cel-go v0.13.0 //v0.6.0
-	github.com/jedib0t/go-pretty/v6 v6.2.2
+	github.com/jedib0t/go-pretty/v6 v6.4.4
 	github.com/matryer/is v1.4.0
 	github.com/mattn/go-sqlite3 v1.14.7
 	google.golang.org/genproto v0.0.0-20221027153422-115e99e71e1c
@@ -14,8 +14,9 @@ require (
 
 require (
 	github.com/antlr/antlr4/runtime/Go/antlr v1.4.10 // indirect
-	github.com/mattn/go-runewidth v0.0.9 // indirect
+	github.com/mattn/go-runewidth v0.0.14 // indirect
+	github.com/rivo/uniseg v0.4.3 // indirect
 	github.com/stoewer/go-strcase v1.2.0 // indirect
-	golang.org/x/sys v0.0.0-20220728004956-3c1f35247d10 // indirect
+	golang.org/x/sys v0.4.0 // indirect
 	golang.org/x/text v0.3.8 // indirect
 )