Upgrade CEL to v0.13.0. 

Evaluation performance decreased, but compilation performance improved:
```
                            │ ./testdata/benchmarks/reference.txt │   ./testdata/benchmarks/after.txt   │
                            │               sec/op                │    sec/op     vs base               │
SimpleRule-2                                          1.014µ ± ∞ ¹   1.163µ ± ∞ ¹  +14.69% (p=0.008 n=5)
SimpleRuleWithDiagnostics-2                           7.484µ ± ∞ ¹   8.191µ ± ∞ ¹   +9.45% (p=0.008 n=5)
RuleWithArray-2                                       1.012µ ± ∞ ¹   1.177µ ± ∞ ¹  +16.30% (p=0.008 n=5)
ProtoWithSelfX-2                                      1.673µ ± ∞ ¹   1.834µ ± ∞ ¹   +9.62% (p=0.008 n=5)
ProtoWithoutSelf-2                                    1.382µ ± ∞ ¹   1.556µ ± ∞ ¹  +12.59% (p=0.008 n=5)
ProtoCreation-2                                       6.220µ ± ∞ ¹   7.102µ ± ∞ ¹  +14.18% (p=0.008 n=5)
Eval2000Rules-2                                       7.530m ± ∞ ¹   7.360m ± ∞ ¹        ~ (p=0.310 n=5)
Eval2000RulesWithSort-2                               7.424m ± ∞ ¹   7.293m ± ∞ ¹        ~ (p=0.056 n=5)
CompileRule-2                                        1256.6µ ± ∞ ¹   765.8µ ± ∞ ¹  -39.06% (p=0.008 n=5)
CompileRuleWithFixedSchema-2                          990.8µ ± ∞ ¹   548.7µ ± ∞ ¹  -44.62% (p=0.008 n=5)
geomean                                               38.78µ         37.24µ         -3.97%
```